+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+- [#3] Make sure we can catch exceptions from `findAll`
+- Make sure callers can easily call a `findAll` variant with indexes
+- Dropped `node@8` support
+
 ### 1.1.1
 
 - [#2] Remove Global Table attributes (`aws:rep:deleting`, `aws:rep:updatetime`, `aws:rep:updateregion`) when creating or updating records.
@@ -14,3 +18,4 @@
 
 [#1]: https://github.com/godaddy/dynastar/pull/1
 [#2]: https://github.com/godaddy/dynastar/pull/2
+[#3]: https://github.com/godaddy/dynastar/pull/3